glfloat

+++ ueb05/src/shader.c
@@ -72,15 +72,15 @@ typedef struct {
 Vertex vert[GRID_LENGTH * GRID_LENGTH * 6];
 
 
-#define GRID_LENGTH_X (2.0)
-#define GRID_LENGTH_Z (M_PI)
+#define GRID_LENGTH_X (GLfloat) (2.0)
+#define GRID_LENGTH_Z (GLfloat) (M_PI)
 
-#define TILE_WIDTH  GRID_LENGTH_Z / (GRID_LENGTH );
-#define TILE_HEIGHT GRID_LENGTH_X / (GRID_LENGTH );
+#define TILE_WIDTH  (GLfloat) (GRID_LENGTH_Z / (GRID_LENGTH ));
+#define TILE_HEIGHT (GLfloat) (GRID_LENGTH_X / (GRID_LENGTH ));
 
 
-#define OFFSET_X (GRID_LENGTH_Z / 2)
-#define OFFSET_Z (GRID_LENGTH_X / 2)
+#define OFFSET_X (GLfloat) (GRID_LENGTH_Z / 2)
+#define OFFSET_Z (GLfloat) (GRID_LENGTH_X / 2)
 
 /**
  * Zeichen-Funktion.
@@ -142,8 +142,8 @@ drawScene(void) {
  */
 void setTextureAndHeight(int idx) {
     for (int i = 0; i < 6; ++i) {
-        vert[idx + i].s = (GLfloat) (vert[idx + i].x + OFFSET_X) / GRID_LENGTH_Z;
-        vert[idx + i].t = (GLfloat) (vert[idx + i].z + OFFSET_Z) / GRID_LENGTH_X;
+        vert[idx + i].s = (vert[idx + i].x + OFFSET_X) / GRID_LENGTH_Z;
+        vert[idx + i].t = (vert[idx + i].z + OFFSET_Z) / GRID_LENGTH_X;
         vert[idx + i].y = 0;
     }
 }
@@ -156,8 +156,8 @@ void setTextureAndHeight(int idx) {
  */
 void initSingleTile(int x, int z, int idx) {
     // Erstes Dreieck
-    vert[idx].x = GRID_LENGTH_Z / GRID_LENGTH * (GLfloat) x - OFFSET_X;
-    vert[idx].z = GRID_LENGTH_X / GRID_LENGTH * (GLfloat) z - OFFSET_Z;
+    vert[idx].x = GRID_LENGTH_Z / GRID_LENGTH * x - OFFSET_X;
+    vert[idx].z = GRID_LENGTH_X / GRID_LENGTH * z - OFFSET_Z;
 
     vert[idx + 1].x = vert[idx].x;
     vert[idx + 1].z = vert[idx].z + TILE_HEIGHT;
